Upgrade yt-dlp and add Instagram support with cookies
- Upgraded yt-dlp from 2024.3.10 to 2025.09.26 for Instagram Reels support
- Added cookies support for Instagram and other platforms requiring auth
- Fixed duration field to accept float values (some videos have fractional durations)
- Added COOKIES_FILE config and docker volume mount
- Removed content_hash unique constraint to allow multiple items without transcripts

Tested:
- ✅ YouTube (transcript extraction)
- ✅ TikTok (audio download)
- ✅ Instagram Reels (audio download with cookies)
- ✅ Cache system working (instant retrieval on repeat URLs)

+++ services/content_extractor_v2/models.py
@@ -15,7 +15,7 @@ class ContentMetadata(BaseModel):
     """Metadata extracted from content."""
     title: str
     channel: Optional[str] = None
-    duration: Optional[int] = None  # seconds
+    duration: Optional[float] = None  # seconds (can be fractional)
     description: Optional[str] = None
     language: Optional[str] = None
     platform: str
